requirements.txt
fastapi
uvicorn[standard]
python-dotenv
supabase_py
openai>=1.14
pydantic
